% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\docType{data}
\name{GITHUB_PERMISSIONS}
\alias{GITHUB_PERMISSIONS}
\title{Known GitHub App permissions}
\format{
An object of class \code{character} of length 29.
}
\usage{
GITHUB_PERMISSIONS
}
\description{
A character vector of all valid GitHub App permission names.
These are used for validation when requesting scoped installation tokens.
}
\keyword{internal}
