% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_jwt_claim.R
\name{generate_jwt_claim}
\alias{generate_jwt_claim}
\title{Generate a JWT claim signed by a private key}
\usage{
generate_jwt_claim(app_id, private_key_path)
}
\arguments{
\item{app_id}{GitHub App ID}

\item{private_key_path}{Path to the GitHub App Private Key}
}
\value{
Signed JWT with claim to the GitHub App
}
\description{
Generate a JWT claim signed by a private key
}
\examples{
\dontrun{
generate_jwt_claim(app_id = "2023", private_key_path = "~/Downloads/key.pem")
}

}
