% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_installation.R
\name{get_first_installation}
\alias{get_first_installation}
\title{Get the first installation of the GitHub App}
\usage{
get_first_installation(jwt)
}
\arguments{
\item{jwt}{Signed JWT with claim to the GitHub App}
}
\value{
A list with all the values of the first GitHub App installation
}
\description{
Get the first installation of the GitHub App
}
\examples{
\dontrun{
get_first_installation(jwt)
}

}
