% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_app_token.R
\name{get_github_app_token}
\alias{get_github_app_token}
\title{Get the GitHub App Installation Token}
\usage{
get_github_app_token(app_id, private_key_path, verbose = TRUE)
}
\arguments{
\item{app_id}{GitHub App ID}

\item{private_key_path}{Path to the GitHub App Private Key}

\item{verbose}{If TRUE (default), will print the token's permissions.
If false, it will only return the token.}
}
\value{
An installation token to make API requests
}
\description{
Get the GitHub App Installation Token
}
\examples{
\dontrun{
get_github_app_token(app_id, private_key_path)
get_github_app_token(app_id, private_key_path, verbose = FALSE)
}

}
