% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_installation_token.R
\name{get_installation_token}
\alias{get_installation_token}
\title{Get the installation token to make API requests}
\usage{
get_installation_token(jwt, installation_id)
}
\arguments{
\item{jwt}{Signed JWT with claim to the GitHub App}

\item{installation_id}{The id of the installation from which we want to obtain the token}
}
\value{
An installation token to make API requests
}
\description{
Get the installation token to make API requests
}
\examples{
\dontrun{
get_installation_token(jwt, installation$id)
}

}
